requirements.txt
Flask
google-generativeai
python-dotenv
requests
beautifulsoup4
selenium
pytest